Major update, added a lot of UEFI components and modules. Currently debugging new Build System.

+++ src/.legacy/kernel/lib32/stdmemory/memory.c
@@ -1,16 +1,16 @@
 #include "kernel/lib32/generic/standard.h"
 
-bool memwithin(void *buffer, size_t buffer_size, void *address, size_t address_size){
-    return ((size_t)address) > ((size_t)buffer) && ((size_t)address) < ((size_t)buffer + buffer_size) &&
-            ((size_t)address + address_size) < ((size_t)buffer + buffer_size);
+bool memwithin(void *buffer, uint64_t buffer_size, void *address, uint64_t address_size){
+    return ((uint64_t)address) > ((uint64_t)buffer) && ((uint64_t)address) < ((uint64_t)buffer + buffer_size) &&
+            ((uint64_t)address + address_size) < ((uint64_t)buffer + buffer_size);
 }
 
-void memset(void *buffer, size_t len, uint8_t val){
-    for(size_t cc = 0; cc < len; ++cc){((uint8_t *)buffer)[cc] = val;}
+void memset(void *buffer, uint64_t len, uint8_t val){
+    for(uint64_t cc = 0; cc < len; ++cc){((uint8_t *)buffer)[cc] = val;}
 }
 
-bool memcmp(void *a, void *b, size_t len){
-    for(size_t cc =0; cc < len; ++cc){
+bool memcmp(void *a, void *b, uint64_t len){
+    for(uint64_t cc =0; cc < len; ++cc){
         if(((uint8_t *)a)[cc] != ((uint8_t *)b)[cc]){
             return false;
         }
@@ -18,15 +18,15 @@ bool memcmp(void *a, void *b, size_t len){
     return true;
 }
 
-void memcpy(void *dst, void *src, size_t len){
-    for(size_t cc =0; cc < len; ++cc){
+void memcpy(void *dst, void *src, uint64_t len){
+    for(uint64_t cc =0; cc < len; ++cc){
         ((uint8_t *)dst)[cc] = ((uint8_t *)src)[cc];
     }
     return;
 }
 
-bool memcheckb(void *a, size_t len, uint8_t b){
-    for(size_t cc =0; cc < len; ++cc){
+bool memcheckb(void *a, uint64_t len, uint8_t b){
+    for(uint64_t cc =0; cc < len; ++cc){
         if(((uint8_t *)a)[cc] != b){
             return false;
         }
@@ -34,8 +34,8 @@ bool memcheckb(void *a, size_t len, uint8_t b){
     return true;
 }
 
-bool memcheckh(void *a, size_t len, u16_t w){
-    for(size_t cc =0; cc < len; ++cc){
+bool memcheckh(void *a, uint64_t len, u16_t w){
+    for(uint64_t cc =0; cc < len; ++cc){
         if(w){
             return false;
         }
@@ -43,8 +43,8 @@ bool memcheckh(void *a, size_t len, u16_t w){
     return true;
 }
 
-bool memcheckl(void *a, size_t len, long l){
-    for(size_t cc =0; cc < len; ++cc){
+bool memcheckl(void *a, uint64_t len, long l){
+    for(uint64_t cc =0; cc < len; ++cc){
         if(((long *)a)[cc] != l){
             return false;
         }
@@ -52,9 +52,9 @@ bool memcheckl(void *a, size_t len, long l){
     return true;
 }
 
-bool memcheckll(void *a, size_t len, size_t z){
+bool memcheckll(void *a, uint64_t len, uint64_t z){
     for(uint32_t cc =0; cc < len; ++cc){
-        if(((size_t *)a)[cc] != z){
+        if(((uint64_t *)a)[cc] != z){
             return false;
         }
     }
